+++ src/app/[apartmentName]/(notAuth)/join/components/scrollBox1.tsx
@@ -1,0 +1,162 @@
+import React from 'react'
+
+export const servise = [
+  {
+    title: '서비스 이용약관 동의',
+    subTitle1: '제 1장 총칙',
+    text1: `제 1조 목적
+    본 약관은 (주)아파트너(이하 '회사'라 함)가 운영하는 인터넷 홈페이지 및 모바일 어플리케이션(이하 ‘아파트너’라 함)에서 제공하는 제반 서비스의 이용과 관련하여 회사와 회원 및 회원 간의 권리, 의무 및 책임사항 기타 필요한 사항을 규정함을 목적으로 합니다.
+    
+    제2조 용어의 정의
+    ① '입주민'은 세대주, 세대원 등 아파트에 거주하는 자 또는, 아파트에 미거주하더라도 아파트를 소유하는 자를 의미합니다.
+    ② '회원'은 아파트너 홈페이지에 회원가입을 한 후, 관리자페이지를 통해 해당 아파트 단지 입주민으로 승인되어 아파트너 서비스를 정식 이용할 수 있는 입주민을 의미합니다.
+    ③ '어플리케이션'은 Android 및 IOS 시스템을 기반으로 하는 스마트폰에서 이용할 수 있는 아파트너 어플리케이션을 의미합니다.
+    ④ '아파트너 홈페이지'는 아파트너가 PC환경에서 이용할 수 있도록, 아파트 단지별로 서비스를 제공하는 웹사이트를 의미합니다.
+    ⑤ '관리자페이지'는 PC환경에서 이용할 수 있는 아파트너가 서비스 제공 중인 단지별 웹사이트를 의미 합니다.
+    ⑥ '아파트너 회사 홈페이지'는 아파트너 전체 서비스에 대한 소개 및 공지 등을 안내하는 아파트너 회사 전용 웹사이트를 의미 합니다.
+    ⑦ “아파트너 유료서비스”라 함은 회사가 서비스의 제공과 관련하여 유료로 제공하는 관련 제반 서비스를 총칭합니다. 아파트너 유료서비스와 관련된 용어의 정의는 이하와 같습니다.
+    1) 유료 서비스 : 회원이 회사에 일정 금액을 지불해야만 이용할 수 있는 회사의 서비스 또는 상품을 의미하거나, 회사 또는 제 3 자와의 거래 내지 약정 조건을 수락하는 대가로 이용하게 되는 회사의 서비스 또는 상품을 의미합니다. 유료서비스의 세부내용은 각각의 서비스 또는 상품 구매 페이지 및 해당 서비스 또는 상품의 결제 페이지에 상세히 설명되어 있으며, 회원은 유료 서비스를 그 설명된 내용을 바탕으로 이용 가능합니다.
+    2) 해지 : 회사 또는 회원이 서비스 개통 후 이용 계약을 장래를 향해 해약하는 것을 말합니다.
+    ⑧ “관리사무소”라 함은 회사와 제휴계약을 체결하고 회원에게 해당 공동주택 관련 각종 관리 등 서비스를 제공하는 공동주택의 관리주체를 의미합니다.
+    제3조 약관의 명시, 효력 및 변경
+    ① 회사는 본 약관의 내용을 회원이 쉽게 접근할 수 있도록 아파트너 회사 홈페이지 또는 어플리케이션에 공지합니다.
+    ② 회사는 필요한 경우 관련 법령을 위배하지 않는 범위 내에서 본 약관을 변경할 수 있습니다. 본 약관이 변경되는 경우 회사는 변경 사항을 시행 일자 7 일 전부터 회원에게 공지 또는 통지하는 것을 원칙으로 하며, 회원에게 불리한 내용으로 변경할 경우에는 그 시행 일자 30 일 전부터 아파트너 회사 홈페이지 또는 어플리케이션 내 공지사항 등을 통해 변경사항을 공지 또는 통지합니다.
+    ③ 회사가 전 항에 따라 공지 또는 통지를 한 후, 회원이 공지 또는 통지일로부터 개정약관 시행일 7 일후까지 거부 의사를 표시하지 아니할 경우 변경된 약관을 승인한 것으로 봅니다. 개정약관에 동의하지 않을 경우 제 11 조 제 1 항에 따라 이용계약을 해지할 수 있습니다.
+    ④ 본 약관은 서비스를 신청한 고객 또는 개인위치정보주체가 본 약관에 동의하고 회사가 제공하는 소정의 절차에 따라 서비스의 이용자로 등록함으로써 효력이 발생합니다.
+    ⑤ 회원이 온라인 또는 스마트폰 앱에서 본 약관의 "동의하기" 버튼을 클릭하였을 경우, 그 회원은 본 약관의 내용을 모두 읽고 이를 충분히 이해하였으며, 그 적용에 동의한 것으로 봅니다.
+    제4조 약관 외 준칙
+    본 약관에서 규정되지 않은 사항에 대해서는 공동주택관리법 및 관련법령에 따르며 회사가 정한 서비스의 이용약관, 운영정책 및 규칙 등 (이하 '세부지침')의 규정을 따릅니다. 또한, 본 약관과 세부지침의 내용이 충돌할 경우 세부지침에 따릅니다.
+    
+    제5조 서비스의 중단
+    ① 회사는 컴퓨터 등 정보통신설비의 보수점검, 교체 및 고장, 통신의 두절 등의 사유가 발생한 경우에는 서비스의 제공을 일시적으로 중단할 수 있습니다.
+    ② 사업 종목의 전환, 사업의 포기, 업체 간의 통합 등의 이유로 서비스를 제공할 수 없게 되는 경우에는 회사는 회원에게 통지하거나, 회원이 알아볼 수 있도록 별도의 공지사항으로 게시합니다.
+    제6조 회원에 대한 통지
+    ① 회사는 이동전화 단문메시지서비스(SMS), 어플리케이션 푸시알림(App push)등으로 회원에게 통지할 수 있습니다.
+    ② 회사는 불특정다수 회원에 대한 통지의 경우 공지사항으로 게시함으로써 개별 통지를 갈음할 수 있습니다.
+    제7조 정보의 제공 및 광고의 게재
+    ① 회사는 광고를 포함한 다양한 정보 제공을 위해 서비스 내 배너, 팝업 등 다양한 방법으로 게재할 수 있습니다. 마케팅 정보 수신 동의를 한 이용자에게 한하여 맞춤형 정보를 제공 받을 수 있습니다.
+    단, 마케팅 정보 수신 동의를 하지 않은 이용자는 맞춤형 정보를 제공받을 수 없습니다.
+    ② 회사는 서비스의 운영과 관련하여 서비스 화면, 홈페이지, 팝업에 대하여 회사가 지정하는 위치에 공지사항 및 광고등을 게재할 수 있습니다.
+    ③ 회사는 서비스 이용 중 필요하다고 인정되는 다양한 마케팅 정보 등을 전자우편, 서신, sms, 전화, 알림 메시지 (push notification) 및 이용자의 연락처와 연동된 sns서비스 등의 방법으로 수신동의를 한 이용자에게 한하여 제공할 수 있습니다. 이 경우 이용자의 통신환경 또는 요금구조에 따라 이용자가 데이터 요금등을 부담할 수 있습니다. 회사는 수신거절을 위한 방법을 이용자에게 제공합니다. 이용자는 회사에서 제공하는 광고성 정보를 수신하지 않기 위해 안내된 수신거부방법에 따라 수신거부 의사를 표현할 수 있습니다. 단, 이미 발송된 광고성 정보에 대해서는 적용하지 않습니다.
+    ④ 회사는 광고성 정보 수신을 거부한 이용자에게 [정보통신망 이용촉진 및 정보보호 등에 관한 법률] 을 준수하여 재수신동의 전까지 광고를 제공하지 않습니다.
+    ⑤ 이용자는 회사가 제공하는 서비스와 관련하여 자료 또는 기타 정보를 변경, 수정, 제한하는 등의 조치를 취하지 않습니다.
+    ⑥ 회사는 서비스 개선 및 이용자 대상 서비스 소개 등을 위한 목적으로 이용자 개인에 대한 추가정보를 요구할 수 있으며, 요청에 대해 이용자는 승낙하여 추가정보를 제공하거나 거부할 수 있습니다.
+    ⑦ 회사는 서비스의 원활하고 안정적인 운영 및 서비스 품질의 개선을 위하여 회원의 개인정보를 제외한 이용자의 모바일 기기 정보(설정, 사양, 운영체제, 버전 등)를 수집, 활용할 수 있습니다.
+    제8조 유료 및 부가서비스 제공을 위한 위탁
+    ① 회사는 회원에게 제공하는 서비스에 대하여 이용요금(이하"서비스 이용료"라 합니다)를 부과할 수 있습니다.
+    1) 직접판매 : 회사가 제공하는 유료서비스를 회원에게 결제 수단을 통해 직접 제공하는 방식을 말합니다.
+    2) 대행판매 : 회사가 제공하는 유료서비스를 제3자를 통해 판매하는 방식으로 해당 판매 상품의 이용 방식은 제휴사가 제공하는 유료서비스 안내에 따릅니다.
+    3) 자동결제 상품: 회원이 1개월 단위로 유료서비스를 갱신하기로 하고, 그 요금 역시 1개월의 정기로 결제하기로 회사에 신청한 경우, 회원이 지정한 결제 일자 및 결제 수단에 따라 유료서비스의 이용 요금이 자동으로 결제되는 상품을 의미합니다.
+    ② 회사가 제공하는 서비스 중 일부는 회사가 계약을 맺은 외부업체 (공동책임업체) 에 의해 위탁될 수 있습니다. 그러한 경우 회원에게 제공되는 부가 또는 유료 서비스에 관한 규정사항은 각 위탁된 외부업체의 약관에 따릅니다.
+    제9조 결제방식
+    ① 유료서비스의 이용에 대한 대금지급방법은 다음 각 호의 방법 중 가능한 방법으로 할 수 있습니다. 다만, 회사는 회원의 지급방법에 대하여 어떠한 명목의 수수료도 추가하여 징수하지 않습니다.
+    1) 폰뱅킹, 인터넷뱅킹, 메일 뱅킹 등의 각종 계좌이체
+    2) 선불카드, 직불카드, 신용카드 등의 각종 카드결제
+    3) 온라인무통장입금
+    4) 전자화폐에 의한 결제
+    5) 마일리지 등 회사가 지급한 포인트에 의한 결제
+    6) 회사와 계약을 맺었거나 회사가 인정한 상품권에 의한 결제
+    7) 전화 또는 휴대전화를 이용한 결제
+    8) 기타 전자적 지급방법에 의한 대금지급 등
+    제10조 자동결제 및 해지
+    ① 회사는 자동결제 갱신 및 이용요금 청구를 위해, 서비스 이용기간 및 서비스 이용기간 종료 후 일정 기간 동안 회원의 결제관련정보를 보유할 수 있습니다.
+    ② 회원이 자동결제 상품을 신청한 후 별도의 해지 신청을 하지 않을 경우, 유료서비스의 실제 이용 여부와 상관없이 매월 지정된 결제 일자에 지정된 결재 수단으로 해당 상품이 자동 결제됩니다.
+    ③ 자동결제 해지로 인한 서비스 만료일은 지정된 결제 일자로부터 1개월 입니다. 자동결제 해지 시 회원은 지정된 결제 일자로부터 1개월 동안 서비스를 이용할 수 있습니다.
+    ④ 타인의 결제 정보를 동의 없이 사용할 경우 관계법령에 의해 민/형사상 책임을 지게 될 수 있습니다.
+    ⑤ 결제 정보 변경, 신용카드 및 휴대전화의 분실, 기타의 사유로 정기결제가 이루어지지 않을 경우 마지막 정기결제일로부터 1개월 이 지난 후 서비스 이용이 자동 중단됩니다.
+    ⑥ 이용요금 미납 등 회원의 귀책사유로 인한 정기결제 중단 및 이에 따른 서비스 이용 중단으로 인한 손해에 대해 회사는 책임을 지지 않습니다.
+    제11조 회원의 청약 철회 및 결제취소
+    ① 회사와 유료서비스의 이용에 관한 계약을 체결한 회원은 회사로부터 수신 확인의 통지를 받은 날로부터 7일 이내에 청약의 철회를 할 수 있습니다.
+    ② 제1항의 청약철회는 회원이 전화, 전자우편 또는 모사전송으로 회사에 그 의사를 표시한 때에 효력이 발생합니다.
+    ③ 회사는 제1항에 따라 회원이 표시한 청약철회의사표시를 수신한 후 지체 없이 이러한 사실을 회원에게 회신합니다.
+    ④ 회원이 이용신청을 철회한 경우, 회사에 해당 결제의 취소를 신청할 수 있습니다. 다만, 각 호의 경우에는 결제 취소가 불가능합니다.
+    1) 회원의 결제 없이 회사가 회원에 대해 지급한 서비스(회사가 시행한 이벤트, 쿠폰이나 보상의 지급 등)에 대한 결제취소 신청인 경우
+    2) 해당 결제 수단에 대해, 카드사 등 결제 수단을 서비스하는 업체에서 지정한 결제 취소 가능 기간을 경과한 경우
+    ⑤ 결제 취소가 불가능한 결제 수단(계좌이체 등)을 이용해서 유료서비스를 결제한 경우, 본 약관 제 12조에 따라 환불처리만 가능합니다.
+    제12조 환불정책
+    ① 회사는 회원이 청약철회의 의사표시를 한 날로부터 또는 회원에게 계약 해지의 의사표시에 대하여 회신한 날로부터 3영업일 이내에, 회원이 유료서비스 이용 시 선택한 결제수단과 동일한 결제수단(예컨대, 회원이 계좌이체로 결제한 경우, 그 계좌로 회사가 환불금을 납입하는 방식임)으로 환불하며, 동일한 결제수단으로 환불이 불가능한 경우 이를 사전에 고지하여야 합니다. 다만, 수납확인이 필요한 결제수단의 경우에는 수납확인일로부터 3영업일 이내에 이를 환불하도록 합니다.
+    ② 유료서비스에 대해 일부 이용 후, 해지하여 환불을 요청하는 경우에는 서비스를 이용한 기간에 해당하는 금액을 차감 후 남은 금액을 환불합니다.`,
+  },
+  {
+    subTitle1: '제2장 이용계약 및 정보 보호',
+    text1: `제13조 서비스 이용 계약 성립
+    ① 본 서비스 이용계약은 서비스를 이용하고자 하는 회원이 약관 제14에 따라 서비스의 이용을 청약하고, 회사가 청약을 검토 후 승낙함으로써 체결됩니다.
+    ② 회사는 다음 각 호의 어느 하나에 해당하는 이용 계약 신청에 대해서는 승낙을 거절할 수 있습니다.
+    1) 공동주택의 관리주체 또는 입주자대표회의가 아닌 경우
+    2) 실명이 아니거나 타인의 명의 또는 정보를 이용한 경우
+    3) 내용을 허위로 기재하거나 이용 계약 신청 요건을 충족하지 못한 경우
+    4) 관련 법령에서 금지하는 행위 또는 입주민의 안전을 저해할 목적으로 서비스를 신청하는 경우
+    5) 위에서 열거한 사유에 준하는 기타 공공질서, 미풍양속에 반하는 행위를 하는 경우
+    ③ 아파트너 서비스 이용 신청 시 제공한 정보는 관련 법령 및 회사의 개인정보처리방침의 적용을 받습니다.
+    ④ "회원"이 유료서비스를 이용하기 위해서는 이 약관에 동의 후 각 서비스에 따른 이용조건에 따라 이용요금을 지급하여야 합니다.
+    제14조 회원가입 및 회원정보의 변경
+    ① 입주민은 어플리케이션 또는 홈페이지에서 회사가 정한 가입 양식에 따라 회원정보를 기재한 후 본 약관의 내용에 동의한다는 의사표시를 함으로써 회원가입을 신청합니다.
+    ② 회원이 가입을 신청하면, 관리자페이지에서 아파트 관리사무를 담당하는 직원이 신청자가 해당 아파트 입주민임을 확인한 이후에 회원가입이 완료되며, 회사가 그 사실을 회원에게 통지하는 시점에 계약이 성립됩니다.
+    ③ 회원은 회원가입 신청 시 기재한 사항이 변경되었을 경우 온라인(어플리케이션, 홈페이지)으로 직접 수정을 하거나 전자우편 및 기타 상당한 방법으로 회사에 그 변경사항을 알려야 합니다.
+    ④ 제3항의 변경사항을 회사에 알리지 않아 발생한 불이익에 대하여 회사는 책임지지 않습니다.
+    ⑤ 회원가입은 반드시 본인의 진정한 정보를 통해서만 가입할 수 있으며 회사는 회원이 등록한 정보에 대하여 확인조치를 할 수 있습니다. 회원은 회사의 확인조치에 대하여 적극적으로 협력하여야 하며, 만일 이를 준수하지 아니할 경우 회사는 회원이 등록한 정보가 부정한 것으로 처리할 수 있습니다.
+    ⑥ 회사는 회원을 등급별로 구분하여 이용에 차등을 둘 수 있습니다.
+    제15조 서비스의 유지/보수
+    회사는 서비스를 365 일, 24 시간 쉬지 않고 제공하기 위하여 최선을 다합니다. 다만, 장비의 유지/보수를 위한 정기, 임시 점검 또는 다른 상당한 이유로 서비스의 제공이 일시 중단될 수 있으며, 이 때에는 미리 서비스 제공화면에 공지합니다. 만약, 회사로서도 예측할 수 없는 이유로 서비스가 중단된 때에는 회사가 상황을 파악하는 즉시 최대한 빠른 시일 내에 서비스를 복구하도록 노력합니다.
+    
+    제16조 이용계약 해지
+    ① 회원은 서비스의 이용을 더 이상 원치 않을 때는 언제든지 아파트너 서비스 내 제공되는 회원탈퇴 메뉴를 이용하여 서비스 회원 이용계약의 해지 신청을 할 수 있으며, 회사는 이를 즉시 처리합니다.
+    ② 회원 이용계약이 해지되더라도 회원이 작성한 게시물, 제 3 자의 게시물에 작성한 댓글 등은 삭제되지 않고 남아있게 됩니다. 다만, 회원의 개인정보는 법령 및 회사의 개인정보 취급방침에 따라 일정 기간 후 삭제됩니다.
+    ③ 제 2 항의 게시물, 댓글 등의 삭제를 원한다면 회원은 회원 이용계약 해지 전 직접 해당 게시물, 댓글을 삭제해야 합니다.
+    제17조 개인정보보호
+    회사는 서비스의 원활한 제공을 위하여 회원이 동의한 목적과 범위 내에서만 개인정보를 수집/이용하며, 개인정보 보호관련 법령에 따라 안전하게 관리합니다.
+    
+    ① 회원은 개인정보처리와 관련한 자세한 내용을 아파트너 홈페이지 또는 어플리케이션 내 ‘개인정보 처리방침’에서 확인하실 수 있습니다.
+    ② 회사가 수집하는 개인정보 및 항목은 서비스 변경이나 회사의 사정에 따라 변경 가능합니다. 이경우 회원에게 변경사항을 약관 제 3조 2항의 방법으로 통지하고 동의를 얻으며, 법령 또는 별도로 동의하지 않은 개인정보를 제 3자에게 제공하지 않으며, 별도의 약관 변경절차는 생략합니다.
+    ③ 모든 회원은 자신의 개인정보를 책임 있게 관리하여 타인이 회원의 개인정보를 부정하게 이용할 수 없도록 해야 합니다.
+    ④ 회사는 회원의 개인정보를 보호하기 위해 [정보통신망 이용촉진 및 정보보호 등에 관한 법률] 상의 개인정보 유효기간제도에 따라 1년간 미 접속한 회원의 개인정보를 파기 또는 분리하여 별도로 저장/관리하며, 회원의 별도 요청이 없는 한 사용이 제한됩니다.`,
+  },
+  {
+    subTitle1: '제 3장 책임',
+    text1: `제18조 회사의 의무
+    회사는 법령과 이 약관이 금지하거나 공서양속에 반하는 행위를 하지 않으며 이 약관이 정하는 바에 따라 지속적이고, 안정적으로 회원에게 서비스를 제공하기 위해 최선을 다합니다.
+    
+    제19조 회원 아이디 및 비밀번호 관리에 대한 의무
+    ① 회원 아이디와 비밀번호에 관한 관리책임은 회원에게 있습니다. 회원은 자신의 아이디 및 비밀번호를 제3자에게 이용하게 해서는 안 됩니다.
+    ② 회원 자신의 아이디 및 비밀번호를 도난당하거나 제 3 자가 사용하고 있음을 인지한 경우에는 지체 없이 회사에 통보하고 회사의 안내가 있는 경우에는 그에 따라야 합니다.
+    제20조 책임의 한계
+    ① 회사는 무료로 제공하는 정보 및 서비스에 관하여 "개인정보처리방침" 또는 관계 법령의 벌칙, 과태료 규정 등 강행규정에 위배되지 않는 한 원칙적으로 손해를 배상할 책임이 없습니다.
+    ② 회사는 천재지변, 불가항력, 서비스용 설비의 보수, 교체, 점검 및 이에 준하는 사항으로 서비스를 제공할 수 없는 경우에 이에 대한 책임이 면제됩니다.
+    ③ 회사는 회원의 귀책사유로 인한 정보 및 서비스 이용의 장애에 관하여는 책임을 지지 않습니다.
+    제 4장 기타
+    제21조 손해배상 등
+    회사는 회사의 과실로 인하여 회원이 손해를 입게 될 경우 본 약관 및 법령에 따라 회원의 손해를 배상합니다. 다만, 회사는 아래와 같은 손해에 대해서는 책임을 부담하지 않습니다. 또한 회사는 법률상 허용되는 한도 내에서 간접 손해, 특별 손해, 결과적 손해, 징계적 손해 및 징벌적 손해에 대한 책임을 부담하지 않습니다.
+    
+    ① 천재지변 또는 이에 준하는 불가항력의 상태에서 발생한 손해
+    ② 회원의 귀책사유로 서비스 이용에 장애가 발생한 경우
+    ③ 제 3 자가 불법적으로 회사의 서버에 접속하거나 서버를 이용함으로써 발생하는 손해
+    ④ 제 3 자가 악성 프로그램을 전송 또는 유포함으로써 발생하는 손해
+    ⑤ 전송된 데이터의 생략, 누락, 파괴 등으로 발생한 손해, 명예훼손 등 제 3 자가 서비스를 이용하는 과정에서 발생된 손해
+    ⑥ 기타 회사의 고의 또는 과실이 없는 사유로 인해 발생한 손해
+    제22조 분쟁의 해결
+    본 약관 또는 서비스는 대한민국법령에 의하여 규정되고 이행됩니다. 서비스 이용과 관련하여 회사와 회원 간에 분쟁이 발생하면 이의 해결을 위해 회사는 성실히 협의할 것입니다. 그럼에도 불구하고 해결되지 않을 경우 민사적 분쟁의 관할을 서울중앙지방법원으로 규정되고 이행됩니다.
+    
+    제23조 효력의 발휘
+    공고일자 : 2024-03-21
+    시행일자 : 2024-04-01`,
+  },
+]
+
+const ScrollBox1 = () => {
+  return (
+    <div className="scrollbar">
+      {servise.map((item, index) => (
+        <div
+          key={index}
+          className="scrollbar-item">
+          <div className="scrollbar-item-title subTitle2">{item?.title}</div>{' '}
+          <br />
+          <div className="body_05">{item.subTitle1}</div> <br />
+          <div className="body_06">{item.text1}</div>
+        </div>
+      ))}
+    </div>
+  )
+}
+
+export default ScrollBox1
