+++ src/app/[apartmentName]/(notAuth)/login/login.scss
@@ -1,4 +1,4 @@
-@import '../../styles/_mixins.module.scss';
+@import '@/styles/_mixins.module.scss';
 
 .commonLayout {
   gap: 50px;
