--- src/app/join/components/step5.tsx
@@ -1,22 +1,0 @@
-import React from 'react'
-
-function Step5() {
-  return (
-    <div className="step-5">
-      <div className="step-5-item">
-        <p className="body_01">회원가입이 완료되었습니다.</p>
-        <p className="body_01">
-          서비스 사용을 위해서 관리사무소의 ‘아파트 인증’이 필요합니다.
-        </p>
-      </div>
-
-      <p className="title1">
-        관리사무소 : <span className="color"> 1600-3123</span> <br />
-        운영시간:<span className="color">09:00~18:00 </span>{' '}
-        <span>(주말/공휴일 휴무) </span>
-      </p>
-    </div>
-  )
-}
-
-export default Step5
